docker/app_to_json/requirements.txt
Flask==1.1.1
Flask-HTTPAuth==4.2.0
lazr.uri==1.0.3
numpy==1.19.4
pandas==1.2.0
pyOpenSSL
Pillow==7.0.0
PyPDF2==1.26.0
Werkzeug==0.16.1
wrapt==1.12.1
zipp==1.0.0
